`sappho-ast`: split out universal expr types into individual mods.
